changed button name from Edit to Edit Name

+++ client/src/components/NewExercisesCard.js
@@ -1,11 +1,23 @@
 import React, { useState } from 'react';
 
 function NewExercisesCard({exercise, deleteExerciseFromRoutine}) {
+    let [count, setCount] = useState([])
+
+    if (!exercise) {
+        return <div>Loading...</div>
+    }
+
+    const addInput = () => {
+        setCount([...count, 1])
+    }
+
+    const renderCount = count.length > 0 && count.map(x => <input />)
 
     return (
         <form>
             <h2>{exercise.name}</h2>
-            <button>Add Set</button>
+            <button onClick={addInput}>Add Set</button>
+            {renderCount}
             <button onClick={() => deleteExerciseFromRoutine(exercise)}>Delete from Routine</button>
         </form>
     )
